1 camera to 2 cameras
Had A LOT of trouble getting the 2D camera running.

Realized that the only way to rotate it w/o being affected by its root component's rotation was to attach it to a SpringArmComp.

Annoying but its gets the job done.

Also I changed the player BP UAsset a little. But I will prob just change it to Tan's one if he changes anything.