Post-Transition Camera Commit 2
Added an option to make trigger follow the 2D collider's size. Made LevelBoxes disappear in half the time of the swap duration.

Still have not fixed the Camera Boom bug. That's something I may need to ask on Unreal Forums bc I understand the problem but don't know the solution.